requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
celery==5.3.4
redis==5.0.1
django-celery-results==2.5.1
channels-redis==4.1.0
asgiref==3.7.2